prepared deprecation of ! commands, restructured discord package
